@@ -44,31 +36,21 @@ dependencies {
     implementation("com.fasterxml.jackson.module:jackson-module-kotlin")
     implementation("org.jetbrains.kotlin:kotlin-reflect")
 
-    implementation("org.zalando:logbook-logstash")
-    implementation("org.zalando:logbook-spring-boot-starter")
-
-    testImplementation("org.springframework:spring-webflux") // for WebTestClient
     testImplementation("org.springframework.boot:spring-boot-starter-test")
-    testImplementation("org.springframework.restdocs:spring-restdocs-mockmvc")
     testImplementation("org.springframework.security:spring-security-test")
 
     testImplementation("com.ninja-squad:springmockk")
-    testImplementation("io.github.logrecorder:logrecorder-assertions")
-    testImplementation("io.github.logrecorder:logrecorder-junit5")
-    testImplementation("io.github.logrecorder:logrecorder-logback")
     testImplementation("io.mockk:mockk-jvm")
-
-    detektPlugins("io.gitlab.arturbosch.detekt:detekt-formatting")
 }
 
 tasks {
     withType<JavaCompile> {
-        sourceCompatibility = "17"
-        targetCompatibility = "17"
+        sourceCompatibility = "21"
+        targetCompatibility = "21"
     }
     withType<KotlinCompile> {
         kotlinOptions {
-            jvmTarget = "17"
+            jvmTarget = "21"
             freeCompilerArgs += "-Xjsr305=strict"
             javaParameters = true
         }
